Extract assert_true helper method
- extend make file to cover all future cardtests to avoid having to touch it again.
- add stubs of cardtests so make can complete without errors
