+++ projects/robbinni/dominion/Makefile
@@ -33,33 +33,57 @@ runtests: testDrawCard
 player: player.c interface.o
 	gcc -o player player.c -g  dominion.o rngs.o interface.o $(CFLAGS)
 
+assert_true.o: assert_true.h assert_true.c dominion.o
+	gcc -c assert_true.c -g $(CFLAGS)
+
 tests: unittestresults.out
 
-unittestresults.out: unittest1.c unittest2.c unittest3.c unittest4.c dominion.o rngs.o
+unittestresults.out: unittest1.c unittest2.c unittest3.c unittest4.c cardtest1.c cardtest2.c cardtest3.c cardtest4.c assert_true.o dominion.o rngs.o
 	echo "Result for running unit tests:" > unittestresults.out
 
 	echo "unittest1.c:" >> unittestresults.out
-	gcc -o unittest1 rngs.o dominion.o unittest1.c $(CFLAGS)
+	gcc -o unittest1 rngs.o dominion.o assert_true.o unittest1.c $(CFLAGS)
 	unittest1 >> unittestresults.out
 	gcov -b -f dominion.c >> unittestresults.out
 
 	echo "unittest2.c:" >> unittestresults.out
-	gcc -o unittest2 rngs.o dominion.o unittest2.c $(CFLAGS)
+	gcc -o unittest2 rngs.o dominion.o assert_true.o unittest2.c $(CFLAGS)
 	unittest2 >> unittestresults.out
 	gcov -b -f dominion.c >> unittestresults.out
 
 	echo "unittest3.c:" >> unittestresults.out
-	gcc -o unittest3 rngs.o dominion.o unittest3.c $(CFLAGS)
+	gcc -o unittest3 rngs.o dominion.o assert_true.o unittest3.c $(CFLAGS)
 	unittest3 >> unittestresults.out
 	gcov -b -f dominion.c >> unittestresults.out
 
 	echo "unittest4.c:" >> unittestresults.out
-	gcc -o unittest4 rngs.o dominion.o unittest4.c $(CFLAGS)
+	gcc -o unittest4 rngs.o dominion.o assert_true.o unittest4.c $(CFLAGS)
 	unittest4 >> unittestresults.out
 	gcov -b -f dominion.c >> unittestresults.out
+
+	echo "cardtest1.c:" >> unittestresults.out
+	gcc -o cardtest1 rngs.o dominion.o assert_true.o cardtest1.c $(CFLAGS)
+	cardtest1 >> unittestresults.out
+	gcov -b -f dominion.c >> unittestresults.out
+
+	echo "cardtest2.c:" >> unittestresults.out
+	gcc -o cardtest2 rngs.o dominion.o assert_true.o cardtest2.c $(CFLAGS)
+	cardtest2 >> unittestresults.out
+	gcov -b -f dominion.c >> unittestresults.out
+
+	echo "cardtest3.c:" >> unittestresults.out
+	gcc -o cardtest3 rngs.o dominion.o assert_true.o cardtest3.c $(CFLAGS)
+	cardtest3 >> unittestresults.out
+	gcov -b -f dominion.c >> unittestresults.out
+
+	echo "cardtest4.c:" >> unittestresults.out
+	gcc -o cardtest4 rngs.o dominion.o assert_true.o cardtest4.c $(CFLAGS)
+	cardtest4 >> unittestresults.out
+	gcov -b -f dominion.c >> unittestresults.out
+
 
 all: playdom player testDrawCard testBuyCard badTestDrawCard
 
 clean:
 	rm -f *.o playdom.exe playdom test.exe test player player.exe testInit testInit.exe *.gcov *.gcda *.gcno *.so
-	rm -f badTestDrawCard testDrawCard unittest1 unittest2 unittest3 unittest4
+	rm -f badTestDrawCard testDrawCard unittest1 unittest2 unittest3 unittest4 cardtest1 cardtest2 cardtest3 cardtest4
