+++ projects/robbinni/dominion/assert_true.h
@@ -1,0 +1,5 @@
+/* asser_true.h
+ * Nigel Robbins
+ */
+
+void assertTrue(int, char*, char*);
