+++ projects/robbinni/dominion/cardtest1.c
@@ -1,0 +1,17 @@
+/* cardtest1.c
+ * Nigel Robbins
+ *
+ */
+
+#include"dominion.h"
+#include"dominion_helpers.h"
+#include"rngs.h"
+#include"assert_true.h"
+#include<stdio.h>
+#include<stdlib.h>
+
+int main() {
+  assertTrue(0, "", "failure");
+  assertTrue(1, "", "");
+  return 0;
+}
