+++ projects/robbinni/dominion/cardtest2.c
@@ -1,0 +1,15 @@
+/* cardtest2.c
+ * Nigel Robbins
+ *
+ */
+
+#include"dominion.h"
+#include"dominion_helpers.h"
+#include"rngs.h"
+#include"assert_true.h"
+#include<stdio.h>
+#include<stdlib.h>
+
+int main() {
+  return 0;
+}
